refactor: replace old macro implementation
