--- crates/qubit-macros/src/macros/mod.rs
@@ -1,3 +1,0 @@
-mod handler;
-
-pub use handler::handler;
